+++ internal/runners/packages/uninstall.go
@@ -43,13 +43,12 @@ func (u *Uninstall) Run(params UninstallRunParams, nsType model.NamespaceType) (
 	}
 
 	return requirements.NewRequirementOperation(u.prime).ExecuteRequirementOperation(
-		params.Package.Name,
-		"",
 		nil,
-		0, // bit-width placeholder that does not apply here
-		bpModel.OperationRemoved,
-		ns,
-		nsTypeV,
-		nil,
+		&requirements.Requirement{
+			Name:          params.Package.Name,
+			Operation:     bpModel.OperationRemoved,
+			Namespace:     ns,
+			NamespaceType: nsTypeV,
+		},
 	)
 }
